Updated sawmill yard & applied new field terrain
Sawmill area 90% completed. Still need more sheds for timber storage and
clamps for logs.

Cliff was extended out to sea to make way for ~3 fields nearby F27.

Fields are to be created next and tested along with moving all hidden
collision planes on top of the cliff to allow usage of the whole area.