Upder-the-hood modifications
A few few small details under the hood have been modified so not a lot
of visual yet.

Clifftop has a few fields laid out, additional hedgerows and walls being
erected.
New muddy dirt tracks to be tested with lack of traction.
terrain has been modified to have dirt show through the mud so it's not
such a block of wet mud